Extend Release Editor section with future plans
Two measures by the Release Editor of 23.05 shall improve the process
for creating release notes: Creating a style guide and developing a
linter.

Feedback for both is needed and wanted. The style guide shall be a
generic description with best practice examples and the linter shall
implement this style with pointing out faulty sections.