Release-Managers: Make ongoing conflicts a criteria
The release manager role is primarily one that is about communication. If
a release manager would be part of an ongoing conflict it could prevent
them from effectively communicating with various maintainers on nixpkgs.

There is also a chance it coudl taint and potentially overshadow the
release process, a risk that should be mitigated early on.