Final-Release: use a more descriptive git tag
Use a more descriptive git tag for the release.

The problem with a `24.05`-style tag is, that they *will* land in peoples
pinned sources - and those people will wonder why they never get any
updates.

This changes hopes to make this less prevalent by adding a more descriptive
name for the actual git tag: It tags the branch-off for the new channel
release, but it is *not* the actual channel.